Rewrote trips to output the different schedules for the different routes.
Since each route (except 102) has a M-F schedule and a Sat and a Sun schedule, I rewrote the trip_id for each route to be the route number plus "W" for weekday, "Sa" for Saturday and "Su" for Sunday.  For example, route 1's trip ID for Saturday is now "1Sa".